requirements.txt
flask
gunicorn
psycopg2-binary
bcrypt
